Getting rid of compiler warnings

+++ src/bbcp_Protocol.C
@@ -924,7 +924,7 @@ void bbcp_Protocol::putCSV(char *Host, char *csFn, char *csVal, int csVsz)
 {                                //1234567890123
    struct iovec iov[] = {{(char *)"Checksum: ", 10},
                         {bbcp_Config.csName,strlen(bbcp_Config.csName)},
-                        {(char *)" ", 1}, {csVal, csVsz},
+                        {(char *)" ", 1}, {csVal, static_cast<std::size_t>(csVsz)},
                         {(char *)" ", 1}, {Host, strlen(Host)},
                         {(char *)":", 1}, {csFn, strlen(csFn)},
                         {(char *)"\n",1}};
